test/#13 - refactor UI tests based on the new componenets
