+++ ui-tests/src/test/java/org/training/steps/CustomerTableSteps.java
@@ -22,50 +22,50 @@ public class CustomerTableSteps extends BaseSteps {
   }
 
   public String getNameHeader() {
-    return homePage.getCustomerTableNameHeader();
+    return homePage.getCustomerTable().getCustomerTableNameHeader();
   }
 
   public String getEmailHeader() {
-    return homePage.getCustomerTableEmailHeader();
+    return homePage.getCustomerTable().getCustomerTableEmailHeader();
   }
 
   public String getPasswordHeader() {
-    return homePage.getCustomerTablePasswordHeader();
+    return homePage.getCustomerTable().getCustomerTablePasswordHeader();
   }
 
   public List<Customer> getAllCustomers() {
-    return homePage.getAllCustomers();
+    return homePage.getCustomerTable().getAllCustomers();
   }
 
   public boolean isRowClickable() {
-    return homePage.isCustomerClickable();
+    return homePage.getCustomerTable().isCustomerClickable();
   }
 
   public CustomerTableSteps clickCustomer() {
-    homePage.clickCustomer();
+    homePage.getCustomerTable().clickCustomer();
     return this;
   }
 
   public CustomerTableSteps clickFirstCustomer() {
-    homePage.clickCustomer(1);
+    homePage.getCustomerTable().clickCustomer(1);
     return this;
   }
 
   public CustomerTableSteps clickSecondCustomer() {
-    homePage.clickCustomer(2);
+    homePage.getCustomerTable().clickCustomer(2);
     return this;
   }
 
   public boolean isSelected() {
-    return homePage.getRowClass().contains("selected");
+    return homePage.getCustomerTable().getRowClass().contains("selected");
   }
 
   public boolean firstIsSelected() {
-    return homePage.getRowClass(1).contains("selected");
+    return homePage.getCustomerTable().getRowClass(1).contains("selected");
   }
 
   public boolean secondIsSelected() {
-    return homePage.getRowClass(2).contains("selected");
+    return homePage.getCustomerTable().getRowClass(2).contains("selected");
   }
 
 }
